Create tests for chat notification code.
We had to move the Firestore refs file into the functions src directory
so that the directory structure stayed in the same shape. Otherwise,
it basically put the ref as the root, then had the rest of the code in
lib/functions/src/test etc.

Additionally, this includes a mock for Firestore which can handle getting
activity data, and getting individual user data. In test mode, we use this
Firestore mock so that we can make sure it is getting called with the
correct parameters.