Utilize arrow functions to maintain this context
We were getting cannot read property "foo" of undefined
errors while running this code in production. This is
because the cloud function took a reference to a class
member function - but by the time it was called, it had
lost its reference to the `this` context of the original
class. By using arrow functions, we maintain the `this`
context so that the class members can call other class
members through `this`.